refactor: 예외 클래스 패키지 구조 변경
- 해당 예외가 발생하는 패키지의 하위 패키지(exception)에 위치하도록 변경합니다.
